armv8-m: Guard portasm.h and FreeRTOS.h declarations
Some declarations in portasm.h and FreeRTOS.h will cause
build errors if it's not understood by the assembler.
For example the 'extern' keyword. To avoid these errors
guard any such code inside a #ifdef #endif block so
the code is included in C files but excluded by the
preprocessor in assembly files.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>